src/main/webapp/app/entities/beer/route/beer-routing-resolve.service.spec.ts
jest.mock('@angular/router');

import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, Router } from '@angular/router';
import { of } from 'rxjs';

import { IBeer, Beer } from '../beer.model';
import { BeerService } from '../service/beer.service';

import { BeerRoutingResolveService } from './beer-routing-resolve.service';

describe('Beer routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: BeerRoutingResolveService;
  let service: BeerService;
  let resultBeer: IBeer | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule],
      providers: [Router, ActivatedRouteSnapshot],
    });
    mockRouter = TestBed.inject(Router);
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRouteSnapshot);
    routingResolveService = TestBed.inject(BeerRoutingResolveService);
    service = TestBed.inject(BeerService);
    resultBeer = undefined;
  });

  describe('resolve', () => {
    it('should return IBeer returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultBeer = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultBeer).toEqual({ id: 123 });
    });

    it('should return new IBeer if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultBeer = result;
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultBeer).toEqual(new Beer());
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse({ body: null as unknown as Beer })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultBeer = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultBeer).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
